+++ include/libp2p/security/noise/crypto/message_patterns.hpp
@@ -30,7 +30,14 @@ namespace libp2p::security::noise {
     MessagePatterns messages;
   };
 
-  HandshakePattern handshakeXX();
+  using MP = MessagePattern;
+
+  const inline HandshakePattern handshakeXX{
+      .name = "XX",
+      .initiatorPreMessages = {},
+      .responderPreMessages = {},
+      .messages = {
+          {MP::E}, {MP::E, MP::DHEE, MP::S, MP::DHES}, {MP::S, MP::DHSE}}};
 
 }  // namespace libp2p::security::noise
