+++ include/libp2p/security/noise/crypto/noise_ccp1305.hpp
@@ -10,9 +10,9 @@
 #include <libp2p/security/noise/crypto/interfaces.hpp>
 
 namespace libp2p::security::noise {
-  class CCP1305Impl : public AEADCipher {
+  class NoiseCCP1305Impl : public AEADCipher {
    public:
-    CCP1305Impl(Key32 key);
+    NoiseCCP1305Impl(Key32 key);
 
     outcome::result<ByteArray> encrypt(gsl::span<const uint8_t> precompiled_out,
                                        uint64_t nonce,
@@ -24,12 +24,12 @@ namespace libp2p::security::noise {
                                        gsl::span<const uint8_t> ciphertext,
                                        gsl::span<const uint8_t> aad) override;
 
-  private:
+   private:
     std::unique_ptr<crypto::chachapoly::ChaCha20Poly1305> ccp_;
   };
 
   class NamedCCPImpl : public NamedAEADCipher {
-  public:
+   public:
     ~NamedCCPImpl() override = default;
 
     std::shared_ptr<AEADCipher> cipher(Key32 key) override;
