+++ include/libp2p/security/noise/crypto/noise_sha256.hpp
@@ -11,9 +11,9 @@
 
 namespace libp2p::security::noise {
 
-  class SHA256 : public NamedHash {
+  class NoiseSHA256HasherImpl : public NamedHasher {
    public:
-    std::shared_ptr<libp2p::crypto::Hash> hash() override;
+    std::shared_ptr<libp2p::crypto::Hasher> hash() override;
 
     std::string hashName() const override;
   };
