+++ src/crypto/hmac_provider/CMakeLists.txt
@@ -5,6 +5,7 @@
 
 libp2p_add_library(p2p_hmac_provider
     hmac_provider_impl.cpp
+    hmac_provider_ctr_impl.cpp
     )
 target_link_libraries(p2p_hmac_provider
     p2p_crypto_error
