+++ src/security/noise/CMakeLists.txt
@@ -14,7 +14,6 @@ libp2p_add_library(p2p_noise
     crypto/noise_dh.cpp
     crypto/noise_sha256.cpp
     crypto/noise_ccp1305.cpp
-    crypto/message_patterns.cpp
     crypto/cipher_suite.cpp
     insecure_rw.cpp
     )
