+++ src/security/noise/crypto/noise_ccp1305.cpp
@@ -7,10 +7,10 @@
 
 namespace libp2p::security::noise {
 
-  CCP1305Impl::CCP1305Impl(Key32 key)
+  NoiseCCP1305Impl::NoiseCCP1305Impl(Key32 key)
       : ccp_{std::make_unique<crypto::chachapoly::ChaCha20Poly1305Impl>(key)} {}
 
-  outcome::result<ByteArray> CCP1305Impl::encrypt(
+  outcome::result<ByteArray> NoiseCCP1305Impl::encrypt(
       gsl::span<const uint8_t> precompiled_out, uint64_t nonce,
       gsl::span<const uint8_t> plaintext, gsl::span<const uint8_t> aad) {
     auto n = ccp_->uint64toNonce(nonce);
@@ -21,7 +21,7 @@ namespace libp2p::security::noise {
     return res;
   }
 
-  outcome::result<ByteArray> CCP1305Impl::decrypt(
+  outcome::result<ByteArray> NoiseCCP1305Impl::decrypt(
       gsl::span<const uint8_t> precompiled_out, uint64_t nonce,
       gsl::span<const uint8_t> ciphertext, gsl::span<const uint8_t> aad) {
     auto n = ccp_->uint64toNonce(nonce);
@@ -33,7 +33,7 @@ namespace libp2p::security::noise {
   }
 
   std::shared_ptr<AEADCipher> NamedCCPImpl::cipher(Key32 key) {
-    return std::make_shared<CCP1305Impl>(key);
+    return std::make_shared<NoiseCCP1305Impl>(key);
   }
 
   std::string NamedCCPImpl::cipherName() const {
